api/requirements.txt
azure-functions
anthropic
scipy
numpy
